refactor: Split search_content
