+++ src/command.rs
@@ -87,33 +87,7 @@ async fn find(db: Synced<impl MeigenDatabase>, opt: FindOptions<'_>) -> Result<S
 
 mod search;
 pub use search::author::search_author;
-
-pub async fn search_content(
-    db: Synced<impl MeigenDatabase>,
-    content: &str,
-    show_count: Option<u8>,
-    page: Option<u32>,
-) -> Result<String> {
-    let page = page.unwrap_or(0);
-    let (show_count, clamp_msg) = option!({
-        value: show_count,
-        default: 5,
-        min: 1,
-        max: 10
-    });
-
-    find(
-        db,
-        FindOptions {
-            author: None,
-            content: Some(content),
-            offset: page * (show_count as u32),
-            limit: show_count,
-        },
-    )
-    .await
-    .edit(|x| x.insert_str(0, clamp_msg))
-}
+pub use search::content::search_content;
 
 pub async fn list(
     db: Synced<impl MeigenDatabase>,
