+++ src/command/search.rs
@@ -1,1 +1,2 @@
 pub mod author;
+pub mod content;
